No progress but I think I can complete things...
So it's a hack really but fuck it, it's probably good enough for me.
Doing this better requires rewriting things with my learnings and
architecting it a lot better, I'll get around to that some day.

For now I just need some basic autocompletion so I can move on, it
doesn't need to scale that much and I can rely on having a fair bit of
memory available.

On cursor idle or post eval (not sure yet) Clojure will be run to find
all symbols and keywords that can be completed. Yep, everything in
memory. EVERYTHING. That massive list will then be sent to Neovim and
set in a variable as an array of strings.

When an omnicomplete request comes through it'll refer to this
pre-populated cache. Hopefully providing instant omnicompletion without
_that_ much overhead.

You could do this better with rpcrequest but I painted myself into a
corner with my shitty handling of threads and communication. I should've
solve notify and request on day one, alas, here we are.

I think this'll work, it's just taken me a few days of bashing my head
against this problem to work out how I can fudge this.

Most of this project is fudging and hacking around stuff to get 90% of
the value with 10% of the effort, I think this sticks to that ethos.