Begin working towards an RPC server
So it opens up on a random port and allows you to send and receive
msgpack RPC messages like Neovim does over stdio. Should support any
number of clients, will have to make it so that you can request the
port.

I don't think it'll work right now so I'll write a test client to check
things with next. #16