Make Conjure log size and direction configurable
It adds 3 new variables, `g:conjure_log_direction`,
`g:conjure_log_size_small` and `g:conjure_log_size_large` to define
the direction and size of the Conjure log window.
Also modifies the size meaning. Before the size was the number of
columns to use, now it is the percentage of the window to use