Improve upserting of log window
So now when we don't _need_ to open it but it's already open we keep it
in sync. But if we don't need it and it's not open then we just update
it in the background without bothering the user.

Cursor stays in the correct position etc even if it's not displayed.