Magically resolve file paths if they don't match
This should be useful for people prepling into a container or remote
machine. If the machine says "the file is at /foo/bar/xyz.clj" but your
source is in ./bar/xyz.clj it should find the file and use that path
instead.

I could walk from the bottom up but that risks more false positives, by
being as specific as possible and getting more and more vague we should
hit the right file if it's inside your CWD somewhere with a similar
parent directory structure.