Simplify relative file root, disable by default
You now turn it on by setting relative_file_root, it's not useful in the
general sense by default. It's something you opt into when you need it
since the default of absolute paths is good enough for almost
everything.

I had more people needing to turn it off than turn it on.