Prevent loading readline config on fennel stdio
This fixes #166. The bug was caused by readline inserting control
characters into the fennel repl. By skipping the user readline config,
they are prevented from modifying the repl in arbitrary dangerous ways.
This is fine, because the readline editing facilities are completely
useless in the context of conjure, since we rely solely on neovim's
editing capabilities.

Note: The user is still able to configure the lua readline library
through their ~/.fennelrc file, for example, to persist their history.