Fix #249, `log-buf?` in `log.fnl`
Previously, conjure would `,enter` even log files for Racket.
This was because the function `log-buf?` in log.fnl used
the match method for strings, and log-buf names contain
'-' (hyphen) symbols, which are metacharacters in lua
pattern matching.

Sanitizing the buffer name by escaping the hyphen before
checking solves the issue.