Auxiliary functions for Swank message parsing
Introduce two new functions to deal with the output coming from swank.

`inner-results` will take a swank output string, and return the inner
values that correspond to the 'stdout' and the 'eval-results'
respectively. This returns it as a whole string.

`parse-separated-list` will take any string and separate out the quoted
components. such as: [[ This "is" not "a \"test\""]] -> ["is" "a \"test\""]