Use new parsing functions for swank output
We then tie the two previous functions together and wrap them in a
`parse-result` function.

If the data does not seem to look like a result of a successful
evaluation, we will still print parts of the message to the debug log,
to provide some feedback with what could have gone wrong.