Use declarative config to position the scroll after jumping to results #302
Also modified the mark setting so we go for the 0th column (first since
the column count is 0-indexed) and to go one line after the last line
count so the cursor is position on the first character of the latest
output.

This means you can then use `%` or ,ee etc to work with the very latest
value without moving your cursor.

Next up is rewriting the mark usage a little to use extmarks.