Replace swank with slynk
Slynk is a modern, maintained fork of swank server, so we will be more
up to date if we support slynk instead. We don't want to increase
maintenance burden on ourselves, so we will remove swank server.

Thanks to it being a fork, this is largely just a find and replace.