Improve client quality
Prior to this change i didn't know how to code. We had some hard coded
values and otherwise, and it was rather messy to try and use.

Now we've extracted the "emacs-rex" (remote execute) into a more
composable function, we can clean up the "eval-and-grab-output" function
and introduce a "connection-info" function, which works as expected