Rework tap code so it supports babashka + only one mapping and no options
I've removed some flexiblity but just tried to stream line things so there's only one thing to think about. I'm also trusting the user to work with the raw atom rather than wrapping it in a function.

I think this is for the best, of course I could be wrong. Thank you so much for all of your work on this @EgorKulbachka!