guile: append stray output to log
This checks to see if there was a result found and if so gathers all
text from after the end of the last prompt until '$n = ...'. If not then
everything from the last prompt up until the next is gathered. In either
scenario the stray text is appended to Conjure's log.
This makes the Guile client more consistent with the built-in Fennel
client.

Also, this uses Fennel's support for multiple cond/body pairs within if
forms thus removing a nested if.